pyproject.toml
[project]
name = "jobscan"
version = "0.1.0"
description = "CLI tool for automating job applications"
readme = "README.md"
authors = [
    { name = "Devon Sampson", email = "devonds@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "playwright>=1.48",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "gspread>=6.1",
    "anthropic>=0.40",
    "weasyprint>=62",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "markdown>=3.5",
    "python-docx>=1.2.0",
]

[project.scripts]
jobscan = "jobscan.cli:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
